move user panel to homepage
